Most recent merge
Pull from here for creating tests, then merge into here.

+++ DesperateHousePi/desperatehousepi/Crust/Relationship.java
@@ -15,7 +15,7 @@ public class Relationship {
 	Date lastMeeting;
 	public LinkedList<String> log;
 	
-	int chemistry;
+	double chemistry;
 	
 	/******************************
 	 * A relationship is a set of contacts that the owner has
@@ -25,7 +25,7 @@ public class Relationship {
 	 * @param new_value - The inital value of the relationship
 	 * @author Michael
 	 ******************************/
-	public Relationship(String new_owner, String contactName, String address, int new_value){
+	public Relationship(String new_owner, String contactName, String address, double new_value){
 		owner = new_owner;
 		nameOfContact = contactName;
 		addressOfContact = address;
@@ -48,14 +48,14 @@ public class Relationship {
 	 * @return The value of this relationship
 	 * @author Michael
 	 */
-	public int getChemistry(){ return chemistry; }
+	public double getChemistry(){ return chemistry; }
 	
 	/*****************************
 	 * Sets the value of the relationship, this is bounded by 100 and -100
 	 * @author Brad and Michael
 	 * @Edited 11/24/13
 	 *****************************/
-	public void setChemistry(int c){
+	public void setChemistry(double c){
 		if(c>100)
 			chemistry = 100;
 		else if(c<-100)
